Conflicts:
	src/error_messages/error_messages.h
	src/os_auth/main-server.c

+++ src/headers/randombytes.h
@@ -3,7 +3,7 @@
 #ifndef __RANDOMBYTES_H
 #define __RANDOMBYTES_H
 
-void randombytes(void *ptr, unsigned int length);
+void randombytes(void *ptr, size_t length);
 void srandom_init(void);
 
 #endif
